refactor downloader function
