+++ m3u/middlewares.go
@@ -2,6 +2,7 @@ package m3u
 
 import (
 	"log"
+	"m3u-stream-merger/utils"
 	"os"
 	"regexp"
 	"strings"
@@ -44,4 +45,26 @@ func groupTitleParser(value string) string {
 
 func tvgLogoParser(value string) string {
 	return generalParser(value)
+}
+
+func checkIncludeGroup(groups []string, line string) bool {
+	debug := os.Getenv("DEBUG") == "true"
+	if debug {
+		utils.SafeLogPrintf(nil, nil, "[DEBUG] Checking if line includes group: %s\n", line)
+	}
+
+	if len(groups) == 0 {
+		return true
+	} else {
+		for _, group := range groups {
+			toMatch := "group-title=" + "\"" + group + "\""
+			if strings.Contains(strings.ToLower(line), toMatch) {
+				if debug {
+					utils.SafeLogPrintf(nil, nil, "[DEBUG] Line matches group: %s\n", group)
+				}
+				return true
+			}
+		}
+		return false
+	}
 }
